feat: unique email and username validation done

+++ cosmo_user/views.py
@@ -45,29 +45,44 @@ def user_register(request):
         return render(request, 'cosmo_user/register.html')
 
     if request.method == 'POST':
-        try:
-            if register_user.register_django_user(request):
-                user = User.objects.get(username=request.POST['username'])
-                
-                if register_user.register_cosmo_user(user=user):
-                    messages.success(request, "You have been successfully register.", extra_tags="1")
-                    return HttpResponseRedirect(reverse('user-index'))
+        if User.objects.filter(email=request.POST['email']).exists():
+            messages.success(request, "User with this email alerady exists.", extra_tags="0")
+            context = {}
+            context.update({'first_name':request.POST['first_name']})
+            context.update({'last_name':request.POST['last_name']})
+            context.update({'username':request.POST['username']})
+            return render(request, "cosmo_user/register.html", context=context)
+        elif User.objects.filter(username=request.POST['username']).exists():
+            messages.success(request, "Please choose different username.", extra_tags="0")
+            context = {}
+            context.update({'first_name':request.POST['first_name']})
+            context.update({'last_name':request.POST['last_name']})
+            context.update({'email':request.POST['email']})
+            return render(request, "cosmo_user/register.html", context=context)
+        else:
+            try:
+                if register_user.register_django_user(request):
+                    user = User.objects.get(username=request.POST['username'])
+                    
+                    if register_user.register_cosmo_user(user=user):
+                        messages.success(request, "You have been successfully register.", extra_tags="1")
+                        return HttpResponseRedirect(reverse('user-index'))
+                    else:
+                        try:
+                            user.delete()
+                            messages.success(request, "Could not register. Please try again.", extra_tags="1")
+                            return HttpResponseRedirect(reverse('user-register'))
+                        except Exception as e:
+                            print(e)
+                            return render(request, 'cosmo_user/register.html')
                 else:
-                    try:
-                        user.delete()
-                        messages.success(request, "Could not register. Please try again.", extra_tags="1")
-                        return HttpResponseRedirect(reverse('user-register'))
-                    except Exception as e:
-                        print(e)
-                        return render(request, 'cosmo_user/register.html')
-            else:
+                    messages.success(request, "Could not register. Please try again.", extra_tags="1")
+                    return HttpResponseRedirect(reverse('user-register'))
+
+            except Exception as e:
+                print(e)
                 messages.success(request, "Could not register. Please try again.", extra_tags="1")
                 return HttpResponseRedirect(reverse('user-register'))
-
-        except Exception as e:
-            print(e)
-            messages.success(request, "Could not register. Please try again.", extra_tags="1")
-            return HttpResponseRedirect(reverse('user-register'))
